Update alcumus-info-tutorial.md
A user configuring the SSO for their company had entered a URL in the Identifier (Entity ID) but had not ticked the checkbox entitled Default.
Not ticking the checkbox resulted in a fault SAML message and failure to authenticate. The checkbox must therefore be ticked so I have made this clear.